Fixes for some warnings I get in Visual Studio 2008

+++ Source/Engine/models.h
@@ -26,7 +26,7 @@ class Model {
 			} else PPA_MATCHES( "image" ) {
 				image.Load( value );
 			} else PPA_MATCHES( "mass" ) {
-				mass = atoi( value.c_str() );
+				mass = (short)atoi( value.c_str() );
 			} else PPA_MATCHES( "rotationsPerSecond" ) {
 				rotPerSecond = static_cast<float>(atof( value.c_str() ));
 			} else PPA_MATCHES( "engine" ) {
@@ -38,9 +38,9 @@ class Model {
 					Log::Error( "Model parser could not find engine '%s'.", value.c_str() );
 				}
 			} else PPA_MATCHES( "thrustOffset" ) {
-				thrustOffset = atoi( value.c_str() );
+				thrustOffset = (short)atoi( value.c_str() );
 			} else PPA_MATCHES( "maxEnergyAbsorption" ) {
-				maxEnergyAbsorption = atoi( value.c_str() );
+				maxEnergyAbsorption = (short)atoi( value.c_str() );
 			}
 			
 			return true;
