+++ Source/Sprites/sprite.h
@@ -56,9 +56,8 @@ class Sprite {
 		Image *GetImage( void ) {
 			return image;
 		}
-		short int GetRadarSize( void ) { return radarSize; }
+		int GetRadarSize( void ) { return radarSize; }
 		Color GetRadarColor( void ) { return radarColor; }
-		
 		virtual int GetDrawOrder( void ) = 0;
 		
 	private:
@@ -70,6 +69,7 @@ class Sprite {
 		float angle;
 		int radarSize;
 		Color radarColor;
+		
 };
 
 #endif // __h_sprite__
